# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aplikasi Gacor API Documentation",
        "description": "Documentation for the Aplikasi Gacor Package Delivery Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/api/pakets": {
            "get": {
                "tags": [
                    "Paket"
                ],
                "summary": "Ambil semua data paket",
                "operationId": "a9384eeed743b5c6291b725cd24e061a",
                "responses": {
                    "200": {
                        "description": "Data paket berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paket retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_paket": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "id_pengirim": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "id_penerima": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "id_laporan": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "berat": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 2.5
                                                    },
                                                    "status_pengiriman": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "jenis": {
                                                        "type": "string",
                                                        "example": "Dokumen"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Paket"
                ],
                "summary": "Buat paket baru",
                "operationId": "ebc2c7cb9a55be96c01761e343bd95da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_pengirim",
                                    "id_penerima",
                                    "id_laporan",
                                    "berat",
                                    "status_pengiriman",
                                    "jenis"
                                ],
                                "properties": {
                                    "id_pengirim": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_penerima": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_laporan": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "berat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.5
                                    },
                                    "status_pengiriman": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "proses",
                                            "dikirim",
                                            "diterima"
                                        ],
                                        "example": "pending"
                                    },
                                    "jenis": {
                                        "type": "string",
                                        "example": "Dokumen"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paket berhasil dibuat"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pakets/{id}": {
            "get": {
                "tags": [
                    "Paket"
                ],
                "summary": "Ambil detail paket",
                "operationId": "9ee06ab35405b885bd239b3925120233",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail paket ditemukan"
                    },
                    "404": {
                        "description": "Paket tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Paket"
                ],
                "summary": "Update data paket",
                "operationId": "c4deeed145699bb1e956e829ff08ffa2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_pengirim": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_penerima": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_laporan": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "berat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3
                                    },
                                    "status_pengiriman": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "proses",
                                            "dikirim",
                                            "diterima"
                                        ],
                                        "example": "proses"
                                    },
                                    "jenis": {
                                        "type": "string",
                                        "example": "Elektronik"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paket berhasil diperbarui"
                    },
                    "404": {
                        "description": "Paket tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Paket"
                ],
                "summary": "Hapus paket",
                "operationId": "8804dc242e5e2fa43d64ec6e16fbe9ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paket berhasil dihapus"
                    },
                    "404": {
                        "description": "Paket tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/paket/filter/status": {
            "get": {
                "tags": [
                    "Paket"
                ],
                "summary": "Filter paket berdasarkan status",
                "operationId": "ccdbe3bc16d5eda1f93c237af2c113e1",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "proses",
                                "dikirim",
                                "diterima"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data paket yang difilter berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/paket/filter/jenis": {
            "get": {
                "tags": [
                    "Paket"
                ],
                "summary": "Filter paket berdasarkan jenis",
                "operationId": "e09996ada1f1365b2c82b60cedf5ff78",
                "parameters": [
                    {
                        "name": "jenis",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data paket yang difilter berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}